#1328 review comments
