+++ karate-core/src/main/java/com/intuit/karate/http/HttpUtils.java
@@ -274,4 +274,24 @@ public class HttpUtils {
         return sb.toString();
     }
 
+    /**
+     * creates a map corresponding to cookie. This is used by appropriate drivers.
+     * Karate Cookie Map is of type <String, String> vs the internal framework expects appropriate types -
+     * for ex boolean for secure and persistent keys.
+     * Solution is to use this utility method or remove those keys from Karate Cookie Map.
+     *
+     * @param karateCookie cookie
+     * @return Map
+     */
+    public static Map<String, Object> convertCookieToActualMap(Cookie karateCookie) {
+        Map<String, Object> cookieMap = new HashMap();
+        cookieMap.put(Cookie.NAME, karateCookie.getName());
+        cookieMap.put(Cookie.VALUE, karateCookie.getValue());
+        cookieMap.put(Cookie.PATH, karateCookie.get(Cookie.PATH));
+        cookieMap.put(Cookie.DOMAIN, karateCookie.get(Cookie.DOMAIN));
+        cookieMap.put(Cookie.SECURE, Boolean.valueOf(karateCookie.get(Cookie.SECURE)));
+        cookieMap.put(Cookie.PERSISTENT, Boolean.valueOf(karateCookie.get(Cookie.PERSISTENT)));
+        return cookieMap;
+    }
+
 }
