+++ karate-demo/src/test/java/driver/demo/Demo01Runner.java
@@ -1,30 +1,18 @@
+
 package driver.demo;
 
 import com.intuit.karate.junit4.Karate;
 import com.intuit.karate.KarateOptions;
 import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
-import test.ServerStart;
 
 @RunWith(Karate.class)
 @KarateOptions(features = "classpath:driver/demo/demo-01.feature")
 public class Demo01Runner {
 
-    private static ServerStart server;
-
-    public static int startServer() throws Exception {
-        if (server == null) { // keep spring boot side alive for all tests including package 'mock'
-            server = new ServerStart();
-            server.start(new String[]{"--server.port=0"}, false);
-        }
-        System.setProperty("demo.server.port", server.getPort() + "");
-        return server.getPort();
-    }
-
     @BeforeClass
-    public static void beforeClass() throws Exception {
+    public static void beforeClass() {
         System.setProperty("karate.env", "mock");
-        startServer();
     }
 
 }
